+++ src/components/Homepage/HomePage.styled.js
@@ -19,8 +19,7 @@ export const InnerWrapper = styled.div`
 export const Box = styled.div`
   display: flex;
   align-items: center;
-  justify-content: center;
-  flex-direction: column;
+  justify-content: space-between;
 `;
 
 export const HeadingText = styled.div`
@@ -34,18 +33,30 @@ export const HeadingText = styled.div`
 export const SubText = styled.span`
   margin: 40px;
   font-family: "Helvetica Neue";
+
+  @media (max-width:991px) {
+    margin: 9px;
+    font-size: 12px;
+  }
 `;
 
 export const SubTextPart = styled.span`
-  font-size: 12rem;
+  font-size: 5rem;
   filter: invert(1) grayscale(1) brightness(2) contrast(2) #fff;
   text-shadow: 0 0 10px rgba(255, 255, 255, 0.8);
+  text-align: start;
+
+   @media (max-width:991px) {
+  font-size: 3rem;
+
+  }
 `;
 
 export const CardBox = styled.ul`
   display: flex;
   flex-wrap: wrap;
   align-items: start;
+  justify-content: flex-start;
   width: 100%;
   padding: 0;
   margin: 0;
