+++ src/db.json
@@ -1,16 +1,28 @@
 {
   "users": [
     {
-      "id": "2",
-      "Date": "2/1/24",
-      "WhatIHaveDone": "i will learn react and there features ",
-      "WhatIwillDo": "i will create route "
+      "id": "68d7",
+      "Date": "2024-01-21",
+      "TaskDone": "fafrfahyyiuy",
+      "PendingTask": "far"
     },
     {
-      "id": "24ee",
-      "Date": "12/12/24",
-      "WhatIHaveDone": "i will learn react ",
-      "WhatIwillDo": "and i will see throug it "
+      "id": "5821",
+      "Date": "2024-02-03",
+      "TaskDone": "react concept and features ",
+      "PendingTask": "noting "
+    },
+    {
+      "id": "b3ad",
+      "Date": "2024-01-31",
+      "TaskDone": " c cb bc ",
+      "PendingTask": "c c bc. bc c c. c"
+    },
+    {
+      "id": "a8b2",
+      "Date": "2024-01-09",
+      "TaskDone": "jojnjnnjbb",
+      "PendingTask": "bhhjb"
     }
   ]
 }
